[MAR-2640] Close replacement review findings
