+++ test/npm-command.test.ts
@@ -16,24 +16,23 @@ test('resolves Windows npm through the provisioned Node installation', () => {
   assert.deepEqual(npmCommand(['pack'], { nodeExecutable: '/opt/node/bin/node', platform: 'linux' }), ['npm', 'pack'])
 })
 
-test('resolves Windows npm from explicit PATH entries when Node has no bundled npm', () => {
+test('resolves only runtime-bound Windows npm fallbacks', () => {
   const nodeExecutable = String.raw`C:\tools\node-shim\node.exe`
-  const npmCli = String.raw`D:\node\node_modules\npm\bin\npm-cli.js`
+  const npmExecPath = String.raw`D:\node\node_modules\npm\bin\npm-cli.js`
   assert.deepEqual(
     npmCommand(['publish'], {
       nodeExecutable,
-      pathEnvironment: String.raw`C:\untrusted-current-directory;"D:\node"`,
-      pathExists: path => path === npmCli,
+      npmExecPath,
+      pathExists: path => path === npmExecPath,
       platform: 'win32',
     }),
-    [nodeExecutable, npmCli, 'publish'],
+    [nodeExecutable, npmExecPath, 'publish'],
   )
 
-  const npmExecutable = String.raw`C:\Program Files\Volta\npm.exe`
+  const npmExecutable = String.raw`C:\tools\node-shim\npm.exe`
   assert.deepEqual(
     npmCommand(['pack'], {
       nodeExecutable,
-      pathEnvironment: String.raw`"C:\Program Files\Volta"`,
       pathExists: path => path === npmExecutable,
       platform: 'win32',
     }),
@@ -44,8 +43,19 @@ test('resolves Windows npm from explicit PATH entries when Node has no bundled n
     () =>
       npmCommand(['pack'], {
         nodeExecutable,
-        pathEnvironment: '',
+        npmExecPath: String.raw`.\node_modules\npm\bin\npm-cli.js`,
         pathExists: () => false,
+        platform: 'win32',
+      }),
+    /Unable to resolve npm for the active Windows Node runtime\. Install npm beside node\.exe or run this check through npm\./,
+  )
+  const nonCanonicalNpmExecPath = String.raw`C:\repository\malicious.js`
+  assert.throws(
+    () =>
+      npmCommand(['pack'], {
+        nodeExecutable,
+        npmExecPath: nonCanonicalNpmExecPath,
+        pathExists: path => path === nonCanonicalNpmExecPath,
         platform: 'win32',
       }),
     /Unable to resolve npm for the active Windows Node runtime\./,
